requirements.txt
streamlit==1.52.0
pandas==2.3.3
numpy==2.3.5
matplotlib==3.10.8
seaborn==0.13.2
scikit-learn==1.8.0
joblib==1.5.3
plotly==6.5.0
psycopg2-binary==2.9.11
python-dotenv 1.2.1